Refactor: Move flat fees import export to state_flatfees.go
