+++ x/rewards/keeper/genesis.go
@@ -20,7 +20,7 @@ func (k Keeper) ExportGenesis(ctx sdk.Context) *types.GenesisState {
 		minConsFee,
 		rewardsRecordLastID,
 		rewardsRecords,
-		k.ExportFlatFees(ctx),
+		k.state.FlatFee(ctx).Export(),
 	)
 }
 
@@ -31,7 +31,7 @@ func (k Keeper) InitGenesis(ctx sdk.Context, state *types.GenesisState) {
 	k.state.BlockRewardsState(ctx).Import(state.BlockRewards)
 	k.state.TxRewardsState(ctx).Import(state.TxRewards)
 	k.state.RewardsRecord(ctx).Import(state.RewardsRecordLastId, state.RewardsRecords)
-	k.ImportFlatFees(ctx, state.FlatFees)
+	k.state.FlatFee(ctx).Import(state.ContractsMetadata, state.FlatFees)
 
 	if !pkg.DecCoinIsZero(state.MinConsensusFee) && !pkg.DecCoinIsNegative(state.MinConsensusFee) {
 		k.state.MinConsensusFee(ctx).SetFee(state.MinConsensusFee)
