+++ x/rewards/keeper/state_flatfee.go
@@ -1,10 +1,13 @@
 package keeper
 
 import (
+	"fmt"
+
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/store/prefix"
 	storeTypes "github.com/cosmos/cosmos-sdk/store/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	sdkErrors "github.com/cosmos/cosmos-sdk/types/errors"
 
 	"github.com/archway-network/archway/x/rewards/types"
 )
@@ -43,4 +46,50 @@ func (s FlatFeeState) GetFee(contractAddr sdk.AccAddress) (sdk.Coin, bool) {
 func (s FlatFeeState) RemoveFee(contractAddr sdk.AccAddress) {
 	store := prefix.NewStore(s.stateStore, types.FlatFeePrefix)
 	store.Delete(contractAddr.Bytes())
+}
+
+// Import initializes state from the genesis flat fees data.
+func (s FlatFeeState) Import(metadata []types.ContractMetadata, flatFees []types.FlatFee) {
+	contractAddrToMetadata := make(map[string]types.ContractMetadata, 0)
+	for _, meta := range metadata {
+		contractAdd := meta.MustGetContractAddress()
+		contractAddrToMetadata[contractAdd.String()] = meta
+	}
+
+	for _, flatFee := range flatFees {
+		contractAddr := flatFee.MustGetContractAddress()
+		fee := flatFee.GetFlatFee()
+
+		if _, ok := contractAddrToMetadata[contractAddr.String()]; !ok {
+			panic(fmt.Sprintf("flat fee: %+v is invalid, err: %s", flatFee, types.ErrContractNotFound))
+		}
+
+		if !fee.Amount.IsPositive() {
+			panic(fmt.Sprintf("flat fee: %+v is invalid, err: %s", flatFee, sdkErrors.ErrInvalidCoins))
+		}
+
+		s.SetFee(contractAddr, fee)
+	}
+}
+
+// Export returns the flat fees genesis data for the state.
+func (s FlatFeeState) Export() []types.FlatFee {
+	store := prefix.NewStore(s.stateStore, types.FlatFeePrefix)
+
+	iterator := store.Iterator(nil, nil)
+	defer iterator.Close()
+
+	var fees = make([]types.FlatFee, 0)
+	for ; iterator.Valid(); iterator.Next() {
+		var coin sdk.Coin
+		contractAddr := sdk.AccAddress(iterator.Key())
+		s.cdc.MustUnmarshal(iterator.Value(), &coin)
+
+		fees = append(fees, types.FlatFee{
+			ContractAddress: contractAddr.String(),
+			FlatFee:         coin,
+		})
+	}
+
+	return fees
 }
